pyproject.toml
[project]
name = "ml-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.3",
    "flask>=3.0.3",
    "werkzeug>=3.0.4",
    "pdfminer>=20191125",
    "openai>=1.45.0",
    "docx>=0.2.4",
    "python-dotenv>=1.0.1",
    "streamlit>=1.38.0",
    "langgraph>=0.2.21",
    "pre-commit>=3.8.0",
    "crewai>=0.55.2",
    "composio-crewai>=0.5.23",
    "e2b>=0.17.2a37",
    "pydantic>=2.9.1",
    "langchain-openai>=0.1.25",
    "together>=1.2.12",
    "watchdog>=4.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/components"]


[tool.black]
line-length = 88

[tool.isort]
profile = "black"
